jupyter-websocket-client: rely on jupyter-serde for kernel spec struct
- Move JupyterKernelspec struct to jupyter-serde crate
- Update imports and references in other crates
- Remove unused fields and dead code annotations
- Add optional 'path' field to KernelLaunchRequest
- Clean up and reorganize struct definitions
