+++ crates/runtimelib/src/jupyter/kernelspec.rs
@@ -6,6 +6,8 @@ use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
 use std::process::Stdio;
 
+use jupyter_serde::JupyterKernelspec;
+
 #[cfg(feature = "tokio-runtime")]
 use tokio::{fs, io::AsyncReadExt, process::Command};
 
@@ -18,20 +20,6 @@ pub struct KernelspecDir {
     pub kernel_name: String,
     pub path: PathBuf,
     pub kernelspec: JupyterKernelspec,
-}
-
-/// Contents of a Jupyter JSON kernelspec file
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct JupyterKernelspec {
-    /// argv must contain `{connection_file}` to be replaced by the client launching the kernel
-    /// For example, `["python3", "-m", "ipykernel_launcher", "-f", "{connection_file}"]`
-    #[serde(default)]
-    pub argv: Vec<String>,
-    pub display_name: String,
-    pub language: String,
-    pub metadata: Option<HashMap<String, Value>>,
-    pub interrupt_mode: Option<String>,
-    pub env: Option<HashMap<String, String>>,
 }
 
 impl KernelspecDir {
